refactor(server): move the llama.cpp-style argument surface into its own module
No behaviour change. `lib.rs` was 9795 lines and the next two changes
both land on the argument surface, so the split comes first, per the
repo rule that a new file beats a new section.

Moved verbatim into `crates/ferrox-server/src/cli.rs`: `ServerArgs`,
`rewrite_llama_style_argv`, `OffloadDevice`, `GpuLayers`,
`print_available_devices`, `cli_bind_addr`, `resolve_hf_repo`,
`apply_cli_overrides`, `BUILT_WITH_METAL`/`BUILT_WITH_CUDA`, and the
four tests that cover them.
